refactor: ORM declaration models in database
