+++ etl_taiga/models/DimProject.py
@@ -1,4 +1,4 @@
-from peewee import *
+from peewee import AutoField, CharField, Model
 from etl_taiga.models import BaseModel
 
 
@@ -7,7 +7,7 @@ class DimProject(BaseModel):
 
     id_project = AutoField(primary_key=True)
     name_project = CharField(max_length=200, null=False)
-    description = CharField(max_length=400, null=False)
+    description = CharField(max_length=9999,null=False)
 
     class Meta:
         table_name = "dim_project"
