+++ etl_taiga/models/DimStatus.py
@@ -1,4 +1,4 @@
-from peewee import *
+from peewee import AutoField, CharField
 from etl_taiga.models import BaseModel
