+++ etl_taiga/models/DimTag.py
@@ -1,4 +1,4 @@
-from peewee import *
+from peewee import AutoField, CharField
 from etl_taiga.models import BaseModel
